+++ js/setup.js
@@ -1,0 +1,3 @@
+const POST_NUMBER = 25;
+
+export { POST_NUMBER };
